Upgrade to Pulumi v2.12.0 and pulumi-terraform-bridge v2.11.0

+++ sdk/python/pulumi_f5bigip/ltm/persistence_profile_dst_addr.py
@@ -5,7 +5,7 @@
 import warnings
 import pulumi
 import pulumi.runtime
-from typing import Any, Dict, List, Mapping, Optional, Tuple, Union
+from typing import Any, Mapping, Optional, Sequence, Union
 from .. import _utilities, _tables
 
 __all__ = ['PersistenceProfileDstAddr']
@@ -25,7 +25,7 @@ class PersistenceProfileDstAddr(pulumi.CustomResource):
                  mirror: Optional[pulumi.Input[str]] = None,
                  name: Optional[pulumi.Input[str]] = None,
                  override_conn_limit: Optional[pulumi.Input[str]] = None,
-                 timeout: Optional[pulumi.Input[float]] = None,
+                 timeout: Optional[pulumi.Input[int]] = None,
                  __props__=None,
                  __name__=None,
                  __opts__=None):
@@ -82,7 +82,7 @@ class PersistenceProfileDstAddr(pulumi.CustomResource):
         :param pulumi.Input[str] name: Name of the persistence profile
         :param pulumi.Input[str] override_conn_limit: To enable _ disable that pool member connection limits are overridden for persisted clients. Per-virtual connection
                limits remain hard limits and are not overridden.
-        :param pulumi.Input[float] timeout: Timeout for persistence of the session
+        :param pulumi.Input[int] timeout: Timeout for persistence of the session
         """
         if __name__ is not None:
             warnings.warn("explicit use of __name__ is deprecated", DeprecationWarning)
@@ -136,7 +136,7 @@ class PersistenceProfileDstAddr(pulumi.CustomResource):
             mirror: Optional[pulumi.Input[str]] = None,
             name: Optional[pulumi.Input[str]] = None,
             override_conn_limit: Optional[pulumi.Input[str]] = None,
-            timeout: Optional[pulumi.Input[float]] = None) -> 'PersistenceProfileDstAddr':
+            timeout: Optional[pulumi.Input[int]] = None) -> 'PersistenceProfileDstAddr':
         """
         Get an existing PersistenceProfileDstAddr resource's state with the given name, id, and optional extra
         properties used to qualify the lookup.
@@ -155,7 +155,7 @@ class PersistenceProfileDstAddr(pulumi.CustomResource):
         :param pulumi.Input[str] name: Name of the persistence profile
         :param pulumi.Input[str] override_conn_limit: To enable _ disable that pool member connection limits are overridden for persisted clients. Per-virtual connection
                limits remain hard limits and are not overridden.
-        :param pulumi.Input[float] timeout: Timeout for persistence of the session
+        :param pulumi.Input[int] timeout: Timeout for persistence of the session
         """
         opts = pulumi.ResourceOptions.merge(opts, pulumi.ResourceOptions(id=id))
 
@@ -255,7 +255,7 @@ class PersistenceProfileDstAddr(pulumi.CustomResource):
 
     @property
     @pulumi.getter
-    def timeout(self) -> pulumi.Output[Optional[float]]:
+    def timeout(self) -> pulumi.Output[Optional[int]]:
         """
         Timeout for persistence of the session
         """
