+++ sdk/python/pulumi_f5bigip/ltm/profile_one_connect.py
@@ -5,7 +5,7 @@
 import warnings
 import pulumi
 import pulumi.runtime
-from typing import Any, Dict, List, Mapping, Optional, Tuple, Union
+from typing import Any, Mapping, Optional, Sequence, Union
 from .. import _utilities, _tables
 
 __all__ = ['ProfileOneConnect']
@@ -17,9 +17,9 @@ class ProfileOneConnect(pulumi.CustomResource):
                  opts: Optional[pulumi.ResourceOptions] = None,
                  defaults_from: Optional[pulumi.Input[str]] = None,
                  idle_timeout_override: Optional[pulumi.Input[str]] = None,
-                 max_age: Optional[pulumi.Input[float]] = None,
-                 max_reuse: Optional[pulumi.Input[float]] = None,
-                 max_size: Optional[pulumi.Input[float]] = None,
+                 max_age: Optional[pulumi.Input[int]] = None,
+                 max_reuse: Optional[pulumi.Input[int]] = None,
+                 max_size: Optional[pulumi.Input[int]] = None,
                  name: Optional[pulumi.Input[str]] = None,
                  partition: Optional[pulumi.Input[str]] = None,
                  share_pools: Optional[pulumi.Input[str]] = None,
@@ -54,9 +54,9 @@ class ProfileOneConnect(pulumi.CustomResource):
         :param pulumi.ResourceOptions opts: Options for the resource.
         :param pulumi.Input[str] defaults_from: Specifies the profile that you want to use as the parent profile. Your new profile inherits all settings and values from the parent profile specified.
         :param pulumi.Input[str] idle_timeout_override: Specifies the number of seconds that a connection is idle before the connection flow is eligible for deletion. Possible values are disabled, indefinite, or a numeric value that you specify. The default value is disabled.
-        :param pulumi.Input[float] max_age: Specifies the maximum age in number of seconds allowed for a connection in the connection reuse pool. For any connection with an age higher than this value, the system removes that connection from the reuse pool. The default value is 86400.
-        :param pulumi.Input[float] max_reuse: Specifies the maximum number of times that a server-side connection can be reused. The default value is 1000.
-        :param pulumi.Input[float] max_size: Specifies the maximum number of connections that the system holds in the connection reuse pool. If the pool is already full, then the server-side connection closes after the response is completed. The default value is 10000.
+        :param pulumi.Input[int] max_age: Specifies the maximum age in number of seconds allowed for a connection in the connection reuse pool. For any connection with an age higher than this value, the system removes that connection from the reuse pool. The default value is 86400.
+        :param pulumi.Input[int] max_reuse: Specifies the maximum number of times that a server-side connection can be reused. The default value is 1000.
+        :param pulumi.Input[int] max_size: Specifies the maximum number of connections that the system holds in the connection reuse pool. If the pool is already full, then the server-side connection closes after the response is completed. The default value is 10000.
         :param pulumi.Input[str] name: Name of the profile_oneconnect
         :param pulumi.Input[str] partition: Displays the administrative partition within which this profile resides
         :param pulumi.Input[str] share_pools: Specify if you want to share the pool, default value is "disabled"
@@ -102,9 +102,9 @@ class ProfileOneConnect(pulumi.CustomResource):
             opts: Optional[pulumi.ResourceOptions] = None,
             defaults_from: Optional[pulumi.Input[str]] = None,
             idle_timeout_override: Optional[pulumi.Input[str]] = None,
-            max_age: Optional[pulumi.Input[float]] = None,
-            max_reuse: Optional[pulumi.Input[float]] = None,
-            max_size: Optional[pulumi.Input[float]] = None,
+            max_age: Optional[pulumi.Input[int]] = None,
+            max_reuse: Optional[pulumi.Input[int]] = None,
+            max_size: Optional[pulumi.Input[int]] = None,
             name: Optional[pulumi.Input[str]] = None,
             partition: Optional[pulumi.Input[str]] = None,
             share_pools: Optional[pulumi.Input[str]] = None,
@@ -118,9 +118,9 @@ class ProfileOneConnect(pulumi.CustomResource):
         :param pulumi.ResourceOptions opts: Options for the resource.
         :param pulumi.Input[str] defaults_from: Specifies the profile that you want to use as the parent profile. Your new profile inherits all settings and values from the parent profile specified.
         :param pulumi.Input[str] idle_timeout_override: Specifies the number of seconds that a connection is idle before the connection flow is eligible for deletion. Possible values are disabled, indefinite, or a numeric value that you specify. The default value is disabled.
-        :param pulumi.Input[float] max_age: Specifies the maximum age in number of seconds allowed for a connection in the connection reuse pool. For any connection with an age higher than this value, the system removes that connection from the reuse pool. The default value is 86400.
-        :param pulumi.Input[float] max_reuse: Specifies the maximum number of times that a server-side connection can be reused. The default value is 1000.
-        :param pulumi.Input[float] max_size: Specifies the maximum number of connections that the system holds in the connection reuse pool. If the pool is already full, then the server-side connection closes after the response is completed. The default value is 10000.
+        :param pulumi.Input[int] max_age: Specifies the maximum age in number of seconds allowed for a connection in the connection reuse pool. For any connection with an age higher than this value, the system removes that connection from the reuse pool. The default value is 86400.
+        :param pulumi.Input[int] max_reuse: Specifies the maximum number of times that a server-side connection can be reused. The default value is 1000.
+        :param pulumi.Input[int] max_size: Specifies the maximum number of connections that the system holds in the connection reuse pool. If the pool is already full, then the server-side connection closes after the response is completed. The default value is 10000.
         :param pulumi.Input[str] name: Name of the profile_oneconnect
         :param pulumi.Input[str] partition: Displays the administrative partition within which this profile resides
         :param pulumi.Input[str] share_pools: Specify if you want to share the pool, default value is "disabled"
@@ -159,7 +159,7 @@ class ProfileOneConnect(pulumi.CustomResource):
 
     @property
     @pulumi.getter(name="maxAge")
-    def max_age(self) -> pulumi.Output[Optional[float]]:
+    def max_age(self) -> pulumi.Output[Optional[int]]:
         """
         Specifies the maximum age in number of seconds allowed for a connection in the connection reuse pool. For any connection with an age higher than this value, the system removes that connection from the reuse pool. The default value is 86400.
         """
@@ -167,7 +167,7 @@ class ProfileOneConnect(pulumi.CustomResource):
 
     @property
     @pulumi.getter(name="maxReuse")
-    def max_reuse(self) -> pulumi.Output[Optional[float]]:
+    def max_reuse(self) -> pulumi.Output[Optional[int]]:
         """
         Specifies the maximum number of times that a server-side connection can be reused. The default value is 1000.
         """
@@ -175,7 +175,7 @@ class ProfileOneConnect(pulumi.CustomResource):
 
     @property
     @pulumi.getter(name="maxSize")
-    def max_size(self) -> pulumi.Output[Optional[float]]:
+    def max_size(self) -> pulumi.Output[Optional[int]]:
         """
         Specifies the maximum number of connections that the system holds in the connection reuse pool. If the pool is already full, then the server-side connection closes after the response is completed. The default value is 10000.
         """
